Add Agent config file mounting and agent cli arg parameter.
Closes #10

agent-configmap now has the data proeprty for the configuration
informaiton set dynamically by the file at the end of agent.configFile.

Test updates to set the basic default of the config file name to
agent.conf which requires a different style of json path specification
to accomodate the internal period in the path name.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>