Add in hostPath backing for the spire-agent disk keyManager.
This commit adds a hostPath volume to the SPIRE agent called
RELEASE-NAME-agent-keymanager which is located at
.agent.keyManager.disk.directory or .agent.hostPath.keyManagerDir.

This commit adds a volumemount to the SPIRE agent POD called
RELEASE-NAME-agent-keymanager which is located at
.agent.keyManager.disk.directory.

The agent config file reflects the value in the volumemount in the
agent.conf:plugins.Keymanager[disk].plugin_data.directory value.

Documentation updates to detail the in-node and in-pod locations,
as well as the addition of the hostPath setting in the KeyManager-Disk
section.

Unit tests to verify the settings react correctly to values.yaml
settings.  Updates to the values.schema.yaml.  No negative testing.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>