Initial submission for agent KeyManager controls. (#14)
* Initial submission for agent KeyManager controls.

Defaults to KeyManager "disk" with directory at "/opt/spire/data/agent/"

Adds keys:
- agent.keyManager.type
  (enum [custom, disk, memory], default disk)

- agent.keyManager.disk.directory
  (string, default "/opt/spire/data/agent/"
   when agent.keyManager.type="disk")

- agent.keyManager.custom.name
  (string, default "customKeyManager"
   when agent.keyManager.type="custom")

- agent.keyManager.custom.cmd (required)
  (string, required
   when agent.keyManager.type="custom")

- agent.keyManager.custom.checksum (required)
  (string, required
   when agent.keyManager.type="custom")

- agent.keyManager.custom.enabled
  (boolean, default true
   when  agent.keyManager.type="custom")

- agent.keyManager.custom.data
  (string, required
   when agent.keyManager.type="custom")

Unit tests for above.

Documentation for above.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>

* Change default config file to match in filename to volume mount.

Previous work had a bug, the command line parameter had a file
name default of *.config when the property being mounted into
a file had a name of *.conf

Signed-off-by: Edwin Buck <edwbuck@gmail.com>

* Fix unit tests to match the new alignment to the agent.conf file.

Requires use of a different json path construct to handle the embedded
period in the key name.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>

* Add in hostPath backing for the spire-agent disk keyManager.

This commit adds a hostPath volume to the SPIRE agent called
RELEASE-NAME-agent-keymanager which is located at
.agent.keyManager.disk.directory or .agent.hostPath.keyManagerDir.

This commit adds a volumemount to the SPIRE agent POD called
RELEASE-NAME-agent-keymanager which is located at
.agent.keyManager.disk.directory.

The agent config file reflects the value in the volumemount in the
agent.conf:plugins.Keymanager[disk].plugin_data.directory value.

Documentation updates to detail the in-node and in-pod locations,
as well as the addition of the hostPath setting in the KeyManager-Disk
section.

Unit tests to verify the settings react correctly to values.yaml
settings.  Updates to the values.schema.yaml.  No negative testing.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>

---------

Signed-off-by: Edwin Buck <edwbuck@gmail.com>